Moved parser logic to init
Increased test coverage
Removed unused test runner
